+++ src/render/renderer_impl.hpp
@@ -39,9 +39,9 @@ public:
     virtual void drawTriangle(const UIDim2& position1, const UIDim2& position2, const UIDim2& position3, const UIBounds& bounds, const Color4& color = Color4()) = 0;
     virtual void drawQuad(const UIDim2& position1, const UIDim2& position2, const UIDim2& position3, const UIDim2& position4, const UIBounds& bounds, const Color4& color = Color4()) = 0;
     virtual void drawRect(const UILayout& layout, const UIBounds& bounds, const Color4& color) = 0;
-    virtual void drawStrokeRect(const UILayout& layout, const UIBounds& bounds, const Color4& color, int stroke, const Color4& strokeColor, const UIStrokeAlignment& strokeAlignment) = 0;
+    virtual void drawStrokeRect(const UILayout& layout, const UIBounds& bounds, const Color4& color, float stroke, const Color4& strokeColor, const UIStrokeAlignment& strokeAlignment) = 0;
     virtual void drawRoundedRect(const UILayout& layout, const UIBounds& bounds, const Color4& color) = 0;
-    virtual void drawRoundedStrokeRect(const UILayout& layout, const UIBounds& bounds, const Color4& color, int stroke, const Color4& strokeColor, const UIStrokeAlignment& strokeAlignment) = 0;
+    virtual void drawRoundedStrokeRect(const UILayout& layout, const UIBounds& bounds, const Color4& color, float stroke, const Color4& strokeColor, const UIStrokeAlignment& strokeAlignment) = 0;
     virtual void drawText(const UIDim2& position, const UIBounds& bounds, const std::string& text, const std::string& path, int size, const Color4& color) = 0;
 
     void enableSubpixel();
